// apps/web/src/data/generated/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigDecimal: { input: string; output: string; }
  BigInt: { input: string; output: string; }
  Bytes: { input: any; output: string; }
  /**
   * 8 bytes signed integer
   *
   */
  Int8: { input: any; output: string; }
  /**
   * A string representation of microseconds UNIX timestamp (16 digits)
   *
   */
  Timestamp: { input: any; output: any; }
};

export type Account = {
  __typename?: 'Account';
  /** Delegate address of the token holder which will participate in votings. Delegates don't need to hold any tokens and can even be the token holder itself. */
  delegate?: Maybe<Delegate>;
  /** An Account is any address that holds any amount of Nouns, the id used is the blockchain address. */
  id: Scalars['ID']['output'];
  /** The Nouns owned by this account */
  nouns: Array<Noun>;
  /** Noun balance of this address expressed as a BigInt normalized value for the Nouns ERC721 Token */
  tokenBalance: Scalars['BigInt']['output'];
  /** Noun balance of this address expressed in the smallest unit of the Nouns ERC721 Token */
  tokenBalanceRaw: Scalars['BigInt']['output'];
  /** Total amount of Nouns ever held by this address expressed as a BigInt normalized value for the Nouns ERC721 Token */
  totalTokensHeld: Scalars['BigInt']['output'];
  /** Total amount of Nouns ever held by this address expressed in the smallest unit of the Nouns ERC721 Token */
  totalTokensHeldRaw: Scalars['BigInt']['output'];
};
// ...
export type Auction = {
  __typename?: 'Auction';
  /** The current highest bid amount */
  amount: Scalars['BigInt']['output'];
  /** The account with the current highest bid */
  bidder?: Maybe<Account>;
  /** The auction bids */
  bids: Array<Bid>;
  /** The time that the auction is scheduled to end */
  endTime: Scalars['BigInt']['output'];
  /** The Noun's ERC721 token id */
  id: Scalars['ID']['output'];
  /** Whether or not the auction is a VRGDA auction */
  isVrgda: Scalars['Boolean']['output'];
  /** The Noun */
  noun: Noun;
  /** Whether or not the auction has been settled */
  settled: Scalars['Boolean']['output'];
  /** The time that the auction was settled */
  settledTimestamp?: Maybe<Scalars['BigInt']['output']>;
  /** The time that the auction started */
  startTime: Scalars['BigInt']['output'];
};
// ...
export type Bid = {
  __typename?: 'Bid';
  /** Bid amount */
  amount: Scalars['BigInt']['output'];
  /** The auction being bid in */
  auction: Auction;
  /** Bidder account */
  bidder?: Maybe<Account>;
  /** Block number of the bid */
  blockNumber: Scalars['BigInt']['output'];
  /** The timestamp of the block the bid is in */
  blockTimestamp: Scalars['BigInt']['output'];
  /** The comment left by the bidder */
  comment?: Maybe<Scalars['String']['output']>;
  /** Bid transaction hash */
  id: Scalars['ID']['output'];
  /** The Noun being bid on */
  noun: Noun;
  /** Index of transaction within block */
  txIndex: Scalars['BigInt']['output'];
};
// ...
export type Noun = {
  __typename?: 'Noun';
  /** The Noun's ERC721 token id */
  id: Scalars['ID']['output'];
  /** The owner of the Noun */
  owner: Account;
  /** The seed used to determine the Noun's traits */
  seed?: Maybe<Seed>;
  /** Historical votes for the Noun */
  votes: Array<Vote>;
};
// ...
export enum ProposalStatus {
  Active = 'ACTIVE',
  Cancelled = 'CANCELLED',
  Executed = 'EXECUTED',
  Pending = 'PENDING',
  Queued = 'QUEUED',
  Vetoed = 'VETOED'
}
// ...
export type Query = {
  __typename?: 'Query';
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  account?: Maybe<Account>;
  accounts: Array<Account>;
  auction?: Maybe<Auction>;
  auctions: Array<Auction>;
  bid?: Maybe<Bid>;
  bids: Array<Bid>;
  delegate?: Maybe<Delegate>;
  delegates: Array<Delegate>;
  delegationEvent?: Maybe<DelegationEvent>;
  delegationEvents: Array<DelegationEvent>;
  dynamicQuorumParams?: Maybe<DynamicQuorumParams>;
  dynamicQuorumParams_collection: Array<DynamicQuorumParams>;
  governance?: Maybe<Governance>;
  governances: Array<Governance>;
  noun?: Maybe<Noun>;
  nouns: Array<Noun>;
  proposal?: Maybe<Proposal>;
  proposals: Array<Proposal>;
  seed?: Maybe<Seed>;
  seeds: Array<Seed>;
  transferEvent?: Maybe<TransferEvent>;
  transferEvents: Array<TransferEvent>;
  vote?: Maybe<Vote>;
  votes: Array<Vote>;
};
// ...
export type Seed = {
  __typename?: 'Seed';
  /** The accessory index */
  accessory: Scalars['BigInt']['output'];
  /** The background index */
  background: Scalars['BigInt']['output'];
  /** The body index */
  body: Scalars['BigInt']['output'];
  /** The glasses index */
  glasses: Scalars['BigInt']['output'];
  /** The head index */
  head: Scalars['BigInt']['output'];
  /** The Noun's ERC721 token id */
  id: Scalars['ID']['output'];
};
// ...
export type AuctionQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type AuctionQuery = { __typename?: 'Query', auction?: { __typename?: 'Auction', id: string, amount: string, startTime: string, endTime: string, settled: boolean, isVrgda: boolean, noun: { __typename?: 'Noun', id: string }, bidder?: { __typename?: 'Account', id: string } | null, bids: Array<{ __typename?: 'Bid', id: string, amount: string, blockTimestamp: string, bidder?: { __typename?: 'Account', id: string } | null }> } | null };

export type CurrentAuctionIdQueryVariables = Exact<{ [key: string]: never; }>;


export type CurrentAuctionIdQuery = { __typename?: 'Query', auctions: Array<{ __typename?: 'Auction', id: string }> };

export type NounSwapProposalsForProposerQueryVariables = Exact<{
  proposerAsString: Scalars['String']['input'];
  proposerAsBytes: Scalars['Bytes']['input'];
}>;


export type NounSwapProposalsForProposerQuery = { __typename?: 'Query', proposals: Array<{ __typename?: 'Proposal', id: string, title: string, description: string, status: ProposalStatus, quorumVotes: string, forVotes: string, againstVotes: string, endBlock: string, startBlock: string }> };

export type AllNounsQueryVariables = Exact<{
  batchSize: Scalars['Int']['input'];
  skip: Scalars['Int']['input'];
}>;


export type AllNounsQuery = { __typename?: 'Query', nouns: Array<{ __typename?: 'Noun', id: string, owner: { __typename?: 'Account', id: string }, seed?: { __typename?: 'Seed', background: string, body: string, accessory: string, head: string, glasses: string } | null }> };

export type NounByIdQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type NounByIdQuery = { __typename?: 'Query', noun?: { __typename?: 'Noun', id: string, owner: { __typename?: 'Account', id: string }, seed?: { __typename?: 'Seed', background: string, body: string, accessory: string, head: string, glasses: string } | null } | null };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: NonNullable<DocumentTypeDecoration<TResult, TVariables>['__apiType']>;
  private value: string;
  public __meta__?: Record<string, any> | undefined;

  constructor(value: string, __meta__?: Record<string, any> | undefined) {
    super(value);
    this.value = value;
    this.__meta__ = __meta__;
  }

  override toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}

export const AuctionDocument = new TypedDocumentString(`
    query Auction($id: ID!) {
  auction(id: $id) {
    id
    noun {
      id
    }
    amount
    startTime
    endTime
    bidder {
      id
    }
    settled
    isVrgda
    bids {
      id
      bidder {
        id
      }
      amount
      blockTimestamp
    }
  }
}
    `) as unknown as TypedDocumentString<AuctionQuery, AuctionQueryVariables>;
export const CurrentAuctionIdDocument = new TypedDocumentString(`
    query CurrentAuctionId {
  auctions(orderBy: startTime, orderDirection: desc, first: 1) {
    id
  }
}
    `) as unknown as TypedDocumentString<CurrentAuctionIdQuery, CurrentAuctionIdQueryVariables>;
export const NounSwapProposalsForProposerDocument = new TypedDocumentString(`
    query NounSwapProposalsForProposer($proposerAsString: String!, $proposerAsBytes: Bytes!) {
  proposals(
    where: {proposer: $proposerAsString, title_contains: "NounSwap"}
    first: 1000
    orderBy: id
    orderDirection: desc
  ) {
    id
    title
    description
    status
    quorumVotes
    forVotes
    againstVotes
    endBlock
    startBlock
  }
}
    `) as unknown as TypedDocumentString<NounSwapProposalsForProposerQuery, NounSwapProposalsForProposerQueryVariables>;
export const AllNounsDocument = new TypedDocumentString(`
    query AllNouns($batchSize: Int!, $skip: Int!) {
  nouns(first: $batchSize, skip: $skip) {
    id
    owner {
      id
    }
    seed {
      background
      body
      accessory
      head
      glasses
    }
  }
}
    `) as unknown as TypedDocumentString<AllNounsQuery, AllNounsQueryVariables>;
export const NounByIdDocument = new TypedDocumentString(`
    query NounById($id: ID!) {
  noun(id: $id) {
    id
    owner {
      id
    }
    seed {
      background
      body
      accessory
      head
      glasses
    }
  }
}
    `) as unknown as TypedDocumentString<NounByIdQuery, NounByIdQueryVariables>;